+++ CHANGELOG.md
@@ -488,6 +488,17 @@
 - `Field::name` attribute has been deprecated. Use `Field::label` instead.
 - `Field::inputHtml()` method has been deprecated. Use `Field::cpInputHtml()` instead.
 
+## 2.1.17 - 2024-05-29
+
+### Changed
+- Ensure that sessions exists when calling `formie/forms/refresh-tokens`.
+
+### Fixed
+- Fixed element integrations update matching logic where matched data is empty.
+- Fixed `populateFormValues` values and dynamic Twig.
+- Fixed an error with Phone number fields and `countryCode`.
+- Fixed an error with Phone number fields and `countryName`.
+
 ## 2.1.16 - 2024-05-27
 
 ### Added